requirements.txt
telethon==1.34.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
psycopg2-binary==2.9.9
schedule==1.2.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
transformers==4.38.2
transliterate==1.10.2
tqdm==4.66.2
pandas==2.2.1
